+++ dimas/src/context.rs
@@ -29,8 +29,11 @@
 //!
 
 // region:		--- modules
+// these ones are only for doc needed
+#[cfg(doc)]
+use crate::agent::Agent;
 #[cfg(feature = "liveliness")]
-use crate::com::liveliness_subscriber::LivelinessSubscriber;
+use crate::com::liveliness::LivelinessSubscriber;
 #[cfg(feature = "publisher")]
 use crate::com::publisher::Publisher;
 #[cfg(feature = "query")]
@@ -40,9 +43,9 @@ use crate::com::queryable::Queryable;
 #[cfg(feature = "subscriber")]
 use crate::com::subscriber::Subscriber;
 use crate::com::{
-	communicator::Communicator, liveliness_subscriber::LivelinessSubscriberBuilder,
-	message::Message, publisher::PublisherBuilder, query::QueryBuilder,
-	queryable::QueryableBuilder, subscriber::SubscriberBuilder,
+	communicator::Communicator, liveliness::LivelinessSubscriberBuilder, message::Message,
+	publisher::PublisherBuilder, query::QueryBuilder, queryable::QueryableBuilder,
+	subscriber::SubscriberBuilder,
 };
 use crate::config::Config;
 use crate::error::{DimasError, Result};
@@ -267,12 +270,7 @@ where
 			.map_err(|_| DimasError::ModifyContext("publishers".into()))?
 			.iter_mut()
 			.for_each(|publisher| {
-				if let Err(reason) = publisher.1.de_init() {
-					error!(
-						"could not de-initialize publisher for {}, reason: {}",
-						publisher.1.key_expr, reason
-					);
-				};
+				publisher.1.de_init();
 			});
 
 		// stop all registered subscribers
@@ -321,8 +319,8 @@ where
 		&self,
 	) -> LivelinessSubscriberBuilder<
 		P,
-		crate::com::liveliness_subscriber::NoPutCallback,
-		crate::com::liveliness_subscriber::Storage<P>,
+		crate::com::liveliness::NoPutCallback,
+		crate::com::liveliness::Storage<P>,
 	> {
 		LivelinessSubscriberBuilder::new(self.prefix().clone())
 			.storage(self.liveliness_subscribers.clone())
@@ -334,8 +332,8 @@ where
 		&self,
 	) -> LivelinessSubscriberBuilder<
 		P,
-		crate::com::liveliness_subscriber::NoPutCallback,
-		crate::com::liveliness_subscriber::NoStorage,
+		crate::com::liveliness::NoPutCallback,
+		crate::com::liveliness::NoStorage,
 	> {
 		LivelinessSubscriberBuilder::new(self.prefix().clone())
 	}
@@ -572,14 +570,6 @@ where
 		self.props
 			.write()
 			.map_err(|_| DimasError::WriteProperties.into())
-	}
-
-	/// Create a [`Publisher`]
-	pub(crate) fn create_publisher<'publisher>(
-		&self,
-		key_expr: &str,
-	) -> Result<zenoh::publication::Publisher<'publisher>> {
-		self.communicator.create_publisher(key_expr)
 	}
 
 	/// Method to do an ad hoc publishing for a `topic`